+++ examples/invites.ts
@@ -17,11 +17,7 @@ const serverJoin = async () => {
 
     for(const i of invites.split(/\s+/g)) {
         const res = await join(i.trim(), token);
-        if(!res) {
-            console.error('Did not join guild %s.', i);
-        } else {
-            console.log('Joined guild %s (%s)!', res.guild.name, res.guild.id);
-        }
+        console.log(res ? `Joined guild ${res.guild.name} (${res.guild.id})!` : `Did not join guild ${i}.`);
     }
 }