+++ examples/resend_email.ts
@@ -1,0 +1,15 @@
+import { confirmation } from '../src/email';
+import prompts = require('prompts');
+
+const Confirm = async () => {
+    const { token } = await prompts({
+        type: 'text',
+        name: 'token',
+        message: 'Discord Token:'
+    });
+
+    const res = await confirmation(token);
+    console.log(res ? 'Re-sent confirmation email' : 'Did not re-send confirmation email');
+}
+
+Confirm();