+++ examples/verify_email.ts
@@ -1,11 +1,11 @@
-import Verify from '../src/email.js';
+import { verify } from '../src/email.js';
 import prompts = require('prompts');
 
 /**
  * Verify it, it's pretty simple.
  * @param {string} url URL from the email, **can be the proxied URL**! 
  */
-const verify = async () => {
+const Verify = async () => {
     const { token, url } = await prompts([
         {
             type: 'text',
@@ -19,8 +19,7 @@ const verify = async () => {
         }
     ]);
 
-    const res = await Verify(url, token);
-    console.log(res);
+    const res = await verify(url, token);
 }
 
-verify();
+Verify();