+++ CHANGELOG.md
@@ -6,3 +6,8 @@
 * Tested and working.
 * Fixed issues with master branch where 2captcha would sometimes return a text result even if json is specified.
 * Replace ``khafradev/ask`` with ``prompts`` as it is more complete and feature-rich.
+
+# v1.0.2
+* Re-send email-confirmation email. (**Untested**)
+* Actual typings now.
+* Determined return types for all functions.